Fixup "when to write out arrow functions"
It's pretty tricky as far as I can tell.  Arrow functions have different semantics
and they're nice sugar mostly but implicit return, etc means we have to be careful
and not assume that the semantics are the same.

Thus we pretty-print out arrow functions in the case where the function has been
written with one body statement only: an explicit return.